requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
Automat @ file:///tmp/build/80754af9/automat_1600298431173/work
bcrypt @ file:///C:/ci/bcrypt_1597936263757/work
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606255208697/work
chardet==3.0.4
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
constantly==15.1.0
cryptography==3.3
cssselect==1.1.0
cycler==0.10.0
Django==3.1.4
django-crispy-forms==1.10.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
hyperlink @ file:///tmp/build/80754af9/hyperlink_1600360282274/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
incremental==17.5.0
iniconfig @ file:///tmp/build/80754af9/iniconfig_1602780191262/work
itemadapter @ file:///tmp/build/80754af9/itemadapter_1605069490748/work
itemloaders==1.0.4
jmespath==0.10.0
kiwisolver==1.3.1
logging==0.4.9.6
lxml @ file:///C:/ci/lxml_1606498900999/work
matplotlib==3.3.3
more-itertools @ file:///tmp/build/80754af9/more-itertools_1605111547926/work
mysqlclient @ file:///C:/Users/jurie/Downloads/Installers/mysqlclient/mysqlclient-1.4.6-cp38-cp38-win_amd64.whl
numpy==1.19.4
packaging @ file:///tmp/build/80754af9/packaging_1606930849755/work
pandas==1.1.5
parsel==1.6.0
Pillow==8.0.1
pluggy==0.13.1
Protego==0.1.16
py @ file:///tmp/build/80754af9/py_1593446248552/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyDispatcher==2.0.5
PyHamcrest @ file:///tmp/build/80754af9/pyhamcrest_1594390921726/work
PyMySQL==0.10.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1606517880428/work
pyparsing==2.4.7
pytest==0.0.0
pytest-runner==5.2
python-dateutil==2.8.1
pytz==2020.4
pywin32==227
queuelib==1.5.0
regex==2020.11.13
requests==2.25.0
Scrapy @ file:///C:/ci/scrapy_1606864804541/work
service-identity @ file:///tmp/build/80754af9/service_identity_1592494375577/work
six @ file:///C:/ci/six_1605187374963/work
sqlparse==0.4.1
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
Twisted==20.3.0
urllib3==1.26.2
w3lib==1.22.0
wincertstore==0.2
zope.interface @ file:///C:/ci/zope.interface_1606940376312/work
